Extend expiration of placek.asc for additional year
Summary:
This change extends the expiration date of the public key contained
within placek.asc by an additional year. This adjustment ensures
continued secure operations and trust in communications or processes
relying on this key, in line with our security policies and operational
requirements.

Technical Details:
The diff indicates an update to the expiration date within the
placek.asc file, which is a PGP public key block. Specifically, the
change is within the ASCII armored representation of the key, suggesting
an extension of its validity. This is crucial for maintaining the
integrity and trustworthiness of the system or communications that
depend on this key for encryption or digital signing. Adjustments to
such keys should be carefully reviewed to ensure they align with
security policies and operational needs, as they directly impact the
cryptographic trust model.

Instructions:
* sudo cryptsetup open /dev/disk/by-uuid/e2948dc0-c2ff-4425-bc1b-27dd2ac6674a secret
* mkdir /home/placek/secret
* sudo mount /dev/mapper/secret /home/placek/secret
* export GNUPGHOME=/home/placek/secret/secret/
* gpg --edit-key $KEYID
* gpg -a --export $KEYID > placek.asc