Add KEYID environment variable for GPG operations
Summary:
This update introduces the KEYID environment variable to the system's
security configuration, aimed at simplifying the usage of GPG commands
related to the user's private key. By setting this variable, we
streamline cryptographic operations, making it easier for users to
manage their security processes.

Technical Details:
The change involves adding a new line to the
modules/security/default.nix file to define the KEYID environment
variable with the value 0xE624230BAE5B5299. This addition is a part of
the config attribute set under home.sessionVariables, indicating that
the variable will be available in user sessions. This improvement is
intended to enhance the user experience with GPG by reducing the need to
remember or look up the key ID during operations, thus making security
practices more efficient and less prone to errors.