Add git absorb command to vim
Changes have been made in the `vimrc` file to include the `gitAbsorb()`
function and define the custom command `Gabsorb`. The new function
allows for executing `:G absorb --base` which integrates the `git
absorb` functionality into Vim. Furthermore, the `default.nix` file has
been updated to include the `git-absorb` package, ensuring that the
necessary dependencies are available for the `git absorb` command in the
Vim editor. This update enhances version control features in the editor,
providing a more seamless and efficient workflow for managing changes.