Redraw Vim window after committing the changes
This commit updates the Vim configuration file 'vimrc' in the 'editor'
module. The function 's:gitAyWayCommitMessage()' has been added,
allowing for a more streamlined way to handle commit messages by
executing the 'cmc' command directly in the buffer and then redrawing
the Vim window for a smoother user experience.

Changes:
- Added the function 's:gitAyWayCommitMessage()' to streamline handling
  of commit messages in Vim.
- Modified the autocmd for FileType gitcommit to call
  's:gitAyWayCommitMessage()' function instead of directly executing
  'cmc'.
- Updated the comments in the vimrc file for clarity and maintenance ease.