Ensure proper packages outputs in order to provide tools
Ensure proper packages outputs in order to provide the `jq` and `curl`
tools. This commit updates the `cmc.nix` file. The changes modify the
usage of `jq` and `curl` by explicitly referencing their paths through
`${pkgs.jq}/bin/jq` and `${pkgs.curl}/bin/curl`, respectively. This
adjustment ensures the correct execution of the commands when the script
is run.