Apply the GitHub issue number in front of the commit title if available
This commit introduces a change to the `cmc.nix` file inside the
`modules/utils` directory. Previously, the script retrieved a message
content according to an API response, but now a conditional statement
has been added to check for the presence of an `issue_no` variable. If
the `issue_no` variable is provided, the script will prepend
`[#issue_no]` to the message content obtained from the API response. If
the `issue_no` is not provided, the script will behave as before, simply
echoing the message content. This modification enhances the commit
messages by automatically adding the issue number when applicable.