Fetch the current GitHub issue title and body
This commit adds a new Vim function `s:currentGitHubIssue()` that
utilizes the `gh` CLI tool to retrieve the title and body of the current
GitHub issue related to the branch you are on. The function first
extracts the issue number from the branch name, then uses the `gh issue
view` command to fetch the title and body in JSON format. The obtained
information is then inserted into the current buffer at the beginning.

An additional key mapping `<leader>t` is created to trigger this
functionality, providing a convenient way to fetch the GitHub issue
details and view them in the editor.