Include Git issue information in the note when the note is empty
This commit adds a new function `s:isBufferEmptyButComments()` to check
if the current buffer is empty except for comments. If the buffer is
indeed empty or contains only comments, the function will propagate the
Git issue information to the note. In the `s:toggleTodoFile()` function,
a condition has been added to call the new function and include the Git
issue in the note if the buffer is empty except for comments. This
enhancement ensures that relevant Git issue data is included in the note
when the buffer is empty, making sure important information is always
captured.