Add functionality to compose and report user stories directly from Vim
Added new functions `composeAnIssue()` and `reportAnIssue()` to
streamline this process. `composeAnIssue()` invokes a command to compose
a user story using the Jarvis tool, while `reportAnIssue()` reads the
title of the story, creates an issue on GitHub with that title, assigns
it to the current user, and opens the issue's URL. Key mappings
`<leader>t` and `<leader>T` have been assigned to `composeAnIssue()` and
`reportAnIssue()` functions respectively, to facilitate quick access to
these features.