Add ability to create or update a Pull Request when pushing commits to a branch
This commit introduces a new function within Vim configuration that
enables pushing commits with the ability to create or update a Pull
Request. When pushing commits using the mapping `<leader>O`, the
function `s:gitPush()` is triggered, issuing a push with the flags
`--force-with-lease` and `--follow-tags`. Additionally, it prompts the
user to decide if a Pull Request should be opened. If confirmed, it
calls the function `oyVeyPullRequestDescription()` to set up the Pull
Request.