Set the server name using the remote URL
This commit updates the default.nix file in the editor module to use the
remote URL to set the server name in the fish shellAliases.editor
configuration. By utilizing the Git remote's URL, the server name is now
dynamically set based on the remote repository, ensuring a more accurate
and adaptable configuration.