Add Elixir scripting language to be readily available for use
In this commit, Elixir has been added to the default.nix file in the
utils module. This addition ensures that Elixir is now included as one
of the scripting languages that can be utilized. By including Elixir,
users will have access to a powerful and functional language for
scripting purposes, expanding the range of options available within the
system.