Allow loading an additional vimrc file with direnv statement
This commit introduces the capability to load an extra vimrc file by
utilizing a direnv statement within the vimrc configuration file. The
implementation includes a section in the vimrc file to check for the
existence of the DIRENV_EXTRA_VIMRC environment variable and sources the
specified file if it is both non-empty and readable. By incorporating
this functionality, users can now easily include an additional vimrc
file as per their requirements when working within the Vim editor,
enhancing the customization options available.