Add useful Nix settings
To enhance the Nix configuration and improve system performance, this
commit adds several useful settings in the configuration.nix file for
the omega machine. The changes focus on enabling automatic garbage
collection, configuring the frequency of garbage collection, setting
specific options for garbage collection, and defining extra options for
Nix.

- Enabled automatic garbage collection in Nix with the `nix.gc.automatic
  = true` setting.
- Set daily garbage collection with `nix.gc.dates = "daily"`, and added
  the option `--delete-older-than 30d` to remove items older than 30
  days.
- Introduced additional Nix options such as enabling experimental
  features like nix-command, flakes, and repl-flake, along with enabling
  auto-optimization of the Nix store.