Adjust the instructions in the Jarvis script regarding commit messages
The commit modifies the instructions in the `jarvis.nix` file under
`modules/utils` to provide less detailed guidance on composing a Git
commit message. Previously, the instructions included a lengthy
description of the process of drafting a commit message, which has been
streamlined to focus on the essential aspects. By adjusting the
instructions, users interacting with the Jarvis tool will now receive
clearer and more concise guidance on writing commit messages.