Refine nnn configuration for enhanced user experience
The modifications in this commit revolve around enhancing the
configuration of the `nnn` program to provide users with a better
experience. The configuration for `nnn` has been updated to include
additional features such as enabling with Nerd Icons, setting specific
bookmarks for quicker access to directories like Documents, Downloads,
and Projects, and adding plugins for functionalities like autojump,
togglex, renamer, suedit, among others.