Apply better colors for zathura
This commit enhances the color scheme for Zathura by adjusting the color
settings in the modules/gui/zathura.nix file. The changes include
updating the default foreground and status bar colors to
config.gui.theme.base07 for improved visibility and aesthetics. By
refining the colors used in Zathura's interface, this modification aims
to provide users with a more visually pleasing reading experience,
making it easier to navigate and interact with the application.