Document the packages added to the utils module
This commit documents the packages added to the utils module, providing
a comprehensive list of tools and functionalities to be included in the
configuration. The modifications in the default.nix file within the
utils module reflect the addition of various packages that serve
different purposes, such as managing remote machines, programming
microcontrollers, handling encrypted volumes, and performing image
manipulation, among others.