Refactor tertius.nix for greater flexibility
In this commit, the tertius.nix file was updated to enhance its
flexibility in usage within the environment by utilizing environment
variables. Specifically, the logic for setting the model and language
parameters was adjusted to make use of environment variables
OPENAI_MODEL and OPENAI_LANGUAGE, respectively. This change allows for
more dynamic customization of the parameters used by the tertius script,
enabling a more versatile and adaptable usage experience.